+++ src/App.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route, Navigate } from "react-router-dom";
 import "./styles/App.css";
-import Login from "./components/Login";
-import SignUp from "./components/SignUp";
+import PublicRoute from "./components/routes/publicRoute";
+import Login from "./components/auth/Login";
+import SignUp from "./components/auth/SignUp";
+import ProtectedRoute from "./components/routes/protectedRoutes";
 import AdminPanel from "./components/Admin";
-import UserDetails from "./components/Userdetail";
-import FundingPage from "./components/ManageFunding";
-import PublicRoute from "./components/publicRoute";
-import ProtectedRoute from "./components/protectedRoutes";
+import UserDetails from "./components/user/Userdetail";
+import FundingPage from "./components/funding/ManageFunding";
 
 function App() {
   return (
